fix: code-review fixes
