+++ src/store/modules/eth-base/eth-base-actions.js
@@ -11,6 +11,9 @@ import {
 } from '../../../lib/pending-transactions'
 import { signTransaction, TransactionFactory } from 'web3-eth-accounts'
 import api from '@/lib/nodes/eth'
+import EthContract from 'web3-eth-contract'
+import { isErc20 } from '@/lib/constants'
+import Erc20 from '../erc20/erc20.abi.json'
 
 /** Interval between attempts to fetch the registered tx details */
 const CHUNK_SIZE = 25
@@ -251,6 +254,46 @@ export default function createActions(config) {
       }
 
       context.commit('areOlderLoading', false)
+    },
+
+    /**
+     * Estimate gas limit for ETH or ERC20 transaction
+     * @param {Object} context - Vuex context
+     * @param {Object} params - Parameters
+     * @param {number} params.amount - Transaction amount
+     * @param {string} params.address - Recipient address
+     * @returns {Promise<number|null>} Estimated gas limit or null if estimation failed
+     */
+    estimateGasLimit: {
+      async handler({ state }, { amount, address }) {
+        try {
+          let transaction
+
+          if (isErc20(state.crypto)) {
+            const contract = new EthContract(Erc20, state.contractAddress)
+            const amountWei = utils.toWhole(amount, state.decimals)
+
+            transaction = {
+              from: state.address,
+              to: state.contractAddress,
+              value: '0x0', // ERC20 transfers don't send ETH
+              data: contract.methods.transfer(address, amountWei).encodeABI()
+            }
+          } else {
+            transaction = {
+              from: state.address,
+              to: address,
+              value: utils.toWei(amount)
+            }
+          }
+
+          const gasLimit = await api.useClient((client) => client.estimateGas(transaction))
+          return Number(gasLimit)
+        } catch (error) {
+          console.warn(`${state.crypto} EstimateGas failed:`, error)
+          return null
+        }
+      }
     }
   }
 }
